update examples to use msg folder
- currently in the template there is a msg dir while in some examples there is not. While this is a uber minor thing it makes it a bit nicer to have it all uniform
