+++ erc20/examples/schema.rs
@@ -4,9 +4,8 @@ use std::path::PathBuf;
 
 use schemars::{schema::RootSchema, schema_for};
 
-use cw_erc20::contract::{
-    AllowanceResponse, BalanceResponse, Constants, HandleMsg, InitMsg, QueryMsg,
-};
+use cw_erc20::contract::{AllowanceResponse, BalanceResponse, Constants};
+use cw_erc20::msg::{HandleMsg, InitMsg, QueryMsg};
 
 fn main() {
     let mut pwd = current_dir().unwrap();
