+++ erc20/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod contract;
+pub mod msg;
 
 #[cfg(test)]
 mod tests;
